requirements.txt
# Core dependencies
pyyaml>=6.0.1
python-dotenv>=1.0.0

# Database
sqlite3

# Geocoding and Maps
geopy>=2.4.0
requests>=2.31.0

# Logging
logging

# Type hints
typing-extensions>=4.7.1

# Testing (optional)
pytest>=7.4.0
pytest-cov>=4.1.0

# Development tools (optional)
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.1
